Forbid useless `%` and allow useful `^`
Add checks for `foo % 0`. Allow the usage of `^` with 1.

Extract tests for division and modulus by zero from
`test_useless_math.py` into a new file as the module is
getting crowded.

Add tests and write changelogs for newly added features.

Closes: #1573
