+++ tests/test_visitors/test_ast/test_operators/test_useless_math.py
@@ -2,7 +2,6 @@ import pytest
 
 from wemake_python_styleguide.violations.consistency import (
     MeaninglessNumberOperationViolation,
-    ZeroDivisionViolation,
 )
 from wemake_python_styleguide.visitors.ast.operators import (
     UselessOperatorsVisitor,
@@ -12,59 +11,11 @@ usage_template = 'constant {0}'
 
 
 @pytest.mark.parametrize('expression', [
-    '/= 0',
-    '/= 0.0',
-    '/= -0',
-    '/= -0.0',
-    '/= -0b0',
-    '/= 0b0',
-    '/= 0x0',
-    '/= -0x0',
-    '/= 0o0',
-    '/= -0o0',
-    '/= 0e0',
-    '/= -0e0',
-    '//= 0',
-
-    '/ 0',
-    '/ 0.0',
-    '/ -0',
-    '/ -0.0',
-    '/ -0b0',
-    '/ 0b0',
-    '/ 0x0',
-    '/ -0x0',
-    '/ 0o0',
-    '/ -0o0',
-    '/ 0e0',
-    '/ -0e0',
-    '// 0',
-
-    '* other / 0',
-    '/ 0 * other',
-])
-def test_zero_div(
-    assert_errors,
-    parse_ast_tree,
-    expression,
-    default_options,
-):
-    """Testing that there is no useless zero div."""
-    tree = parse_ast_tree(usage_template.format(expression))
-
-    visitor = UselessOperatorsVisitor(default_options, tree=tree)
-    visitor.run()
-
-    assert_errors(visitor, [ZeroDivisionViolation])
-
-
-@pytest.mark.parametrize('expression', [
     # Math ops:
     '*= 0',
     '**= -0.0',
     '+= 0e0',
     '-= -0b0',
-    '%= 0',
 
     '* 0',
     '** -0.0',
@@ -80,6 +31,7 @@ def test_zero_div(
     '* 0b1',
     '** 1',
     '/ 1.0',
+    '% 0o1',
 
     '* other / 1.0',
     '* 1 * other',
@@ -103,54 +55,13 @@ def test_meaningless_math(
     expression,
     default_options,
 ):
-    """Testing that there is no useless zero div."""
+    """Testing that meaningless number operations are forbidden."""
     tree = parse_ast_tree(usage_template.format(expression))
 
     visitor = UselessOperatorsVisitor(default_options, tree=tree)
     visitor.run()
 
     assert_errors(visitor, [MeaninglessNumberOperationViolation])
-
-
-@pytest.mark.parametrize('expression', [
-    # Div, but not with zero:
-    '/= 0.1',
-    '/= -1.0',
-    '/= -10',
-    '/= -100.0',
-    '/= 0b101',
-    '/= -1e0',
-    '//= 9',
-
-    '/ 10',
-    '/ 0.01',
-    '/ -0.2',
-    '/ -0x11',
-
-    # variables:
-    '* var',
-    '- var',
-    '/ var',
-    '+ var',
-    '*= var',
-    '-= var',
-    '/= var',
-    '+= var',
-    '%= var',
-])
-def test_correct_zero_div(
-    assert_errors,
-    parse_ast_tree,
-    expression,
-    default_options,
-):
-    """Testing that there is no useless zero div."""
-    tree = parse_ast_tree(usage_template.format(expression))
-
-    visitor = UselessOperatorsVisitor(default_options, tree=tree)
-    visitor.run()
-
-    assert_errors(visitor, [])
 
 
 @pytest.mark.parametrize('expression', [
@@ -176,12 +87,12 @@ def test_correct_zero_div(
     '<< 1',
     '| 0b1',
     '^ 2',
+    '^= 0x1',
     '& -1',
 
     '>>= 10',
     '<<= 1',
     '|= 0b1',
-    '^= 0x1',
     '&= -0o1',
 ])
 def test_useful_math(
@@ -190,7 +101,7 @@ def test_useful_math(
     expression,
     default_options,
 ):
-    """Testing that there is no useless zero div."""
+    """Testing that useful math operations are allowed."""
     tree = parse_ast_tree(usage_template.format(expression))
 
     visitor = UselessOperatorsVisitor(default_options, tree=tree)
